# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegra",
    "description": "Production-ready Agent Protocol server",
    "version": "0.17.0"
  },
  "paths": {
...
    "/threads/{thread_id}/state": {
      "get": {
        "tags": [
          "Threads"
        ],
        "summary": "Get Thread State",
        "description": "Get the current state of a thread.\n\nReturns the latest checkpoint's values, pending next nodes, interrupt\ndata, and metadata. If the thread has no associated graph yet (no runs\nexecuted), returns an empty state.",
        "operationId": "get_thread_state_threads__thread_id__state_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "subgraphs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include states from subgraphs",
              "default": false,
              "title": "Subgraphs"
            },
            "description": "Include states from subgraphs"
          },
          {
            "name": "checkpoint_ns",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Checkpoint namespace to scope lookup",
              "title": "Checkpoint Ns"
            },
            "description": "Checkpoint namespace to scope lookup"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadState"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Threads"
        ],
        "summary": "Update Thread State",
        "description": "Update thread state or retrieve it via POST.\n\nWhen `values` is provided, creates a new checkpoint with the updated state.\nUse `as_node` to attribute the update to a specific graph node. When\n`values` is null AND `as_node` is not provided, this endpoint acts as a\nPOST-based alternative to the GET state endpoint (useful when passing\ncomplex checkpoint/subgraph parameters in the request body).\n\nWhen `values` is null AND `as_node` is provided (e.g. ``as_node=\"__copy__\"``\nas LangGraph Studio sends for \"Re-run from here\"), this creates a new\ncheckpoint derived from the supplied ``checkpoint_id`` without applying\nany state change \u2014 used to anchor a subsequent run as a fork of that\ncheckpoint rather than of the thread's latest state.",
        "operationId": "update_thread_state_threads__thread_id__state_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadStateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ThreadState"
                    },
                    {
                      "$ref": "#/components/schemas/ThreadStateUpdateResponse"
                    }
                  ],
                  "title": "Response Update Thread State Threads  Thread Id  State Post"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/threads/{thread_id}/state/{checkpoint_id}": {
      "get": {
        "tags": [
          "Threads"
        ],
        "summary": "Get Thread State At Checkpoint",
        "description": "Get the thread state at a specific checkpoint.\n\nUse this to inspect historical state at any point in the thread's\nexecution history. Returns 404 if the checkpoint does not exist.",
        "operationId": "get_thread_state_at_checkpoint_threads__thread_id__state__checkpoint_id__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "checkpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkpoint Id"
            }
          },
          {
            "name": "subgraphs",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include states from subgraphs",
              "default": false,
              "title": "Subgraphs"
            },
            "description": "Include states from subgraphs"
          },
          {
            "name": "checkpoint_ns",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Checkpoint namespace to scope lookup",
              "title": "Checkpoint Ns"
            },
            "description": "Checkpoint namespace to scope lookup"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadState"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/threads/{thread_id}/runs": {
      "post": {
        "tags": [
          "Thread Runs"
        ],
        "summary": "Create Run",
        "description": "Create and execute a new run.\n\nStarts graph execution asynchronously and returns the run record\nimmediately with status `pending`. Poll the run or use the stream\nendpoint to follow progress. Provide either `input` or `command` (for\nhuman-in-the-loop resumption) but not both.",
        "operationId": "create_run_threads__thread_id__runs_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Thread Runs"
        ],
        "summary": "List Runs",
        "description": "List runs for a thread.\n\nReturns runs ordered by creation time (newest first). Use `status` to\nfilter and `limit`/`offset` to paginate; a full page sets the\n`X-Pagination-Next` cursor header. Pass `select` to return only the listed\nfields.\n\nDeclared without a response model because `select` makes the row shape\ndynamic \u2014 full entities when omitted, projected dicts when given.",
        "operationId": "list_runs_threads__thread_id__runs_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum rows per page.",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum rows per page."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Rows to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Rows to skip."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by run status.",
              "title": "Status"
            },
            "description": "Filter by run status."
          },
          {
            "name": "select",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "run_id",
                      "thread_id",
                      "assistant_id",
                      "created_at",
                      "updated_at",
                      "status",
                      "metadata",
                      "kwargs",
                      "multitask_strategy"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only the listed fields; omit for the full entity.",
              "title": "Select"
            },
            "description": "Return only the listed fields; omit for the full entity."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runs/wait": {
      "post": {
        "tags": [
          "Stateless Runs"
        ],
        "summary": "Stateless Wait For Run",
        "description": "Create a stateless run and wait for completion.\n\nGenerates an ephemeral thread, delegates to the threaded ``wait_for_run``\nendpoint, and deletes the thread after the response finishes streaming\n(unless ``on_completion=\"keep\"``).",
        "operationId": "stateless_wait_for_run_runs_wait_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/stream": {
      "post": {
        "tags": [
          "Stateless Runs"
        ],
        "summary": "Stateless Stream Run",
        "description": "Create a stateless run and stream its execution.\n\nGenerates an ephemeral thread, delegates to the threaded\n``create_and_stream_run`` endpoint, and deletes the thread after the\nstream finishes (unless ``on_completion=\"keep\"``).",
        "operationId": "stateless_stream_run_runs_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server-Sent Events stream",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentProtocolError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "AgentProtocolError",
        "description": "Standard Agent Protocol error response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ThreadStateUpdateResponse": {
        "properties": {
          "checkpoint": {
            "additionalProperties": true,
            "type": "object",
            "title": "Checkpoint",
            "description": "The checkpoint that was created/updated"
          }
        },
        "type": "object",
        "required": [
          "checkpoint"
        ],
        "title": "ThreadStateUpdateResponse",
        "description": "Response model for thread state update"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}